apk: cache APKINDEX in memory
This will probably come back to bite me, but these tend to be small
enough (and few enough) that caching one per URL in memory is going to
be not terrible and speed up searching the index by quite a bit. This is
a lazier approach than caching it in a parsed form, but this supports
both short-form style rendering and full-form style rendering where we
want to show the entire document.

For the short-form, we should probably also cache a parsed/indexed
version of this to make it even faster, but the current size of the
APKINDEXes that I'm dealing with and the roundtrip latency to the
service make this kinda moot for now.