apk: Cache parsed APKINDEX for short form
This is much faster than parsing each time, and generally people aren't
looking at the APKINDEX, so that being slower is kinda fine with me. We
also don't need to cache allt he bytes of the APKINDEX, just the stuff
we need for searching to work. This makes search feel ~instant without a
huge impact on anything else.