apk: Stream APK rendering
This is a first pass at getting APKs to render while we parse them.
Prior to this, large APKs would just hang and show nothing until we had
downloaded and indexed the entire thing. This change adds an additional
Files() method to the layerFS files, which our http fork will use
instead of ReadDir. Files() is implemented using an iterator, so we
yield results as we're Next()ing through the tar reader.

This also uses streaming HTML to indicate that a file is Loading...
until we can replace the slot with how long it took to actually index
the file. While we're indexing, we still render results to the page.

We should make sure that we only index things once by deduping any
in-flight index scans so that when people click e.g. .PKGINFO we don't
kick off a second indexer.