Add error suggestions
Adds an option error wrapper that will output raw unescaped HTML. This
can be used to attach optional suggestions for common errors / next
steps we expect people to take.

For the first impl, adds a suggestion to prompt people to go to the
catalog view if the repository doesn't exist. Also removes the special
casing for cgr.dev/chainguard as a result.